requirements.txt
langchain-community==0.3.2
pinecone-client==2.2.4
python-dotenv==1.0.0
streamlit==1.29.0